Fix bug : Possible segmentation fault with an invalid pointer.
With the old signatures of *LX_Device::gamepadToString()*
and *LX_Gamepad::toString()* if the C-string has a length that is
smaller than the length of the string generated by *LX_Device::gamepadToString()*,
then a buffer overflow will happen.
So, it was possible to write in an area that was out of the range of the C-string,
and consequently, get access to invalid data.
This bug was due to a bad usage of string through C and C++ together.