LX_Window : Added a new constructor and refactored the window creation.
I defined a new signature of the constructor (LX_Window) in order to simplify
the creation of a window.
A window can be created using *LX_WindowInfo* that defines a structure.
This structure contains information about the window that will be created by
the user.
It is possible to give default information to the strucure using one of these
following functions :

 - LX_initWindowInfo	: it defines the default values of the window.
 - LX_loadWindowConfig	: it defines some values from the configuration file.

Author : Luxon JEAN-PIERRE <luxon.jean.pierre@gmail.com>