LX_Window : Refactored the call of *create_window()*.
The private function *create_window()* took an std::string in argument.
It was a problem because every public function (constructors of *LX_Window*)
send an std::string into this function instead of a C-style string.

So, to avoid a copy between the public functions and *create_window()*,
I replaced the string itself with a reference to the string in order to
avoid a useless copy and optimize the function call.

Author : Luxon JEAN-PIERRE <luxon.jean.pierre@gmail.com>