Implemented LX_TextInput.
*LX_TextInput* is a class that handles text input. The text input is activated
at class intantiation and inactivated at class destruction.
This class can handle Unicode string in UTF-8 thanks to UTF8string.

The API provides a callback (*LX_RedrawCallback*) defined as a function object.
This callback was designed to redraw the current window that displays
the user-typed text. It is not necessary to handle text input events,
because *LX_TextInput* does that properly.

So, in order to use LX_TextInput, you need to:
  - Define a class that inherits the operator **()** from *LX_RedrawCallback*.
  - Declare an object of type *LX_TextInput*.
  - Call the function **eventLoop** giving the callback in argument.

Example of code :

    // Define and implement a subclass of LX_RedrawCallback
    class FuncDraw : public LX_RedrawCallback
	{
	 public:

		...
		void operator ()(UTF8String& u8str)
		{
			// Do something
		}
		...
	}

	...

	// In a function that uses the callback
	...
	void textInput()
	{
		FuncDraw callbck;
		LX_TextInput input;

		input.eventLoop(callbck);
    }

Author : Luxon JEAN-PIERRE <luxon.jean.pierre@gmail.com>