LX_Text: Refactored the input.
A part of the code of the text input has been put in a separated function
It was necessary to do that in order to to avoid a redundancy.
This function is used in the function (*paste()*) that pastes
the string from the clipboard into the current buffer at a good position,
that is to say the potision of the cursor.

--