Fixed an issue in the fullscreen mode (linux version).
In some linux systems, the resolution of a window in fullscreen mode
should be different from the resolution before the window have been toggled
in fullscreen (LX_GRAPHICS_FULLSCREEN).

For example, if the window has the following resolution (800 * 600),
this resolution might not be kept in fullscreen mode. So the application should
be rendered at desktop resolution (1600 * 900 on my laptop).

This commit fixed that using the "logical size". In fullscreen mode,
the resolution of the picture is not different but is stretched to fill
the full screen.

--