LX_Physics: Refactored the collision detection between polygons.
In the first version, the collision detection between two polygons was done
checking each intersection between a segment and the others.
This method is too inefficient because it runs in O(n²) time.

I fixed that implementing a new detection collision based on the use of the
Axis-Aligned Bounding Box (AABB) of each polygon and the Separating Axis Theorem (SAT).
Indeed, I had to handle three cases:

- The two polygons are convex and have at least 4 sides.
- One of the two polygons is not convex or has 3 sides.
- One of the two polygons has less than 3 sides (error case → invalid argument)

The first case needs to be solved using the SAT. So this operation can be done
in O(n) time.
The second case has just been solved using the first version of the collision detection.
I did not find a good algorithm that solves this case in O(n) time.

The use of the AABB is necessary to avoid big calculation with huge polygons
(> 10000 sides). The big calculus is only done when the basic collision is detected.

The third case involves digons and monogons. I do not deal with that.
Moreover, If a polygon has no edge, the physics module has to launch an exception.

--