LX_Window: Improved fillCircle.
LX_Window::fillCircle is more efficient.
Actually, this algorithm is based on Bresenham's circle algorithm.
But instead of drawing a lot of points from the center to the circle arc,
this function draws four ranges of lines from 8 specific points in order
to get the circle filled:

    --

        (x_center - x, y_center + x) x---------x (x_center + y, y_center + x)

    (x_center - x, y_center + y) x-----------------x (x_center + x, y_center + y)

										  O

    (x_center - x, y_center - y) x-----------------x (x_center + x, y_center - y)

        (x_center - y, y_center - x) x---------x (x_center + y, y_center - x)

	--
O(x_center, y_center) is the center of the circle.
x and y are computed by the function.

References:
 - Bresenham's circle algorithm/Midpoint circle algorithm:
   http://www.wikiwand.com/en/Midpoint_circle_algorithm

 - Fast algorithm for drawing circles:
   http://stackoverflow.com/questions/1201200/fast-algorithm-for-drawing-filled-circles

 - Midpoint circle algorithm for filled circles (explanation of the first question):
   http://stackoverflow.com/a/10878576

--