LX_Particle: Particle system improvement.
LX_Particle and LX_ParticleSystem hav been hardly improved.

In the old system you had to create particles and add their graphics resource
using setTexture or setSurface.
Moreover, displaying a particle was a bit tricky so that the particle system
has to retrieve the window from the window manager, get graphics data from
the particle and put that on the window.

Now, that is over \(^.^)/!!!!

You just need to provide the LX_Sprite object to the particle at creation
instead of doing that seperately.
The particle system has been simplified in order to easily draw the particles
without getting information about graphics data of each particle
(it does not need that).

--