LX_Mixer: removed loadMusic() and loadSample()
It is dificult to know in a source code if load*() returns an allocated
music/chunk without taking a look on the documentation.
So I prefer to remove them in order to make the user get the music/chunk doing:

	using namespace LX_Mixer;

	...

    LX_Music *m = new LX_Music("some_mus.ogg");
	LX_Chunk *c = new LX_Chunk("some_chunk.wav");

	... // do something

	delete m
	delete c;

Instead of:

    LX_Music *m = loadMusic("some_mus.ogg");	// Impossible to know if it is allocated
	LX_Chunk *c = loadSample("some_chunk.wav");	// unlees you read the documentation

	... // do something

	delete m
	delete c;

**NB**:

A signature of loadSample() is still here.
In this signature the function takes an LX_FileBuffer object to read.
So, it is still possible to do:

    LX_FileBuffer buffer("some_chunk.wav");
	LX_Chunk *c = loadSample(buffer);		// ↔ c = buffer.loadSample();

	... // do something
	delete c

--