LX_Event: defined pushUserEvent()
*pushUserEvent()* allows the application to push user-defined events
onto the event queue.

The internal structure of the user event (LX_UserInput) is exactly identical to *SDL_UserEvent*
(*LX_UserEvent* is an alias of *SDL_UserEvent*).

The main difference is that the user just needs to use these following fields (if necessary):

- code: the user-defined event code
- data1: first data to store
- data2: second data to store

The other fields are just reserved for private uses.

The type of each user-defined event is set to *SDL_USEREVENT*. So, in order to retieve a user event
from *LX_Event::pollEvent()* or *LX_Event::waitEvent()*, it is necessary to add a new case
in the switch instruction:

    LX_Input ev;
	LX_Event::LX_UserInput uev;

	uev.code = 42;
	uev.data1 = some_data;
	uev.data2 = some_other_data;

	pushUserEvent(uev);

	while(pollEvent(ev))
	{
		switch(ev.type)
		{
			case USEREVENT:
				/* Do something */
				break;
		}
	}

Links:

- User event (SDL2): https://wiki.libsdl.org/SDL_UserEvent
- Semantic of the user event (SDL2): http://gamedev.stackexchange.com/questions/112260/how-do-i-correctly-use-sdl-registerevent-to-create-sdl-userevents#115642

--