LX_Event: drag and drop
It is possible to drop a file in a window.

When a file is dropped in a window, the application can get the path of the file
and execute operations on it (read/write, ...).

As of the version of SDL2 (2.0.3) used by the library,
it is not possible to know the window where the file has been dropped in.

--