LX_Library: fix bug and refactoring
There was a bug when the program checks if the gamepad subsystem has been
activated in the library. This bug allows the program to add the set of
gamepad mappings from *config/gamecontrollerdb.txt*, even if the gamepad subsystem
is not activated. → fixed

When the SDL system was initialized with no flags ( → SDL_Init(0) )
the initialization is still working. It was not expected, because normally
LunatiX should refuse the initialization. → fixed

--