LX_FileBuffer: Get a part of a file in the buffer.
Before: LX_FileBuffer loaded the entire file and stored it in
the file buffer.

Now: It is also possible to put a part of a file in the buffer, setting
the offset and the size of data (in bytes) to read.

This features is very useful, in particular for getting data from
a compressed file, or getting an image from an audio file.

--