LX_Mixer: refactoring | UTF-8 issue
The internal member (filename) of LX_Mixer and LX_Chunk is
an std::string instead of UTF8string.

Some systems (Windows, for example) encode the name of their files
in another format (UTF-16/32, ...). This format cannot be compatible with UTF-8.

std::string is implemented as a byte sequence, so this class is less restrictive
than UTF8string. UTF8string is a sequence of codepoints.
So, it needs to check the validity of the string. Otherwise, it throws an exception

Links:

Wikipedia page - https://en.wikipedia.org/wiki/UTF-8
UTF8string page - https://github.com/Gumichan01/utf8_string

--