lunatix: physics/graphical — strict separation (I)
The architecture of the library (physics/graphical parts) is bad.

> Why?

It is bad because I have not separated the graphical part (image position/rectangle)
and the logical part (position, rectangle, line, segment).

Consequently, It is very very hard for the user to provide and use information
about how to draw an image, and making calculation related to physics.
Those are two different things.

> Now

In the previous version I used element related to physics in order to do
graphical operations.

In the incoming version, The user will use graphical elements to do
graphical operations. This is exactly the same thing with physics elements/operations.

--