lunatix: physics/graphical — strict separation (II)
LX_ImgRect contains information about the image that is used to be drawn
on the window. It contains information about:

- the position of the image
- the dimenwion of the image

This structure is completely related to graphics.
It cannot be used in the physics module.

In order to use the physics module, the user must use *LX_FloatPosition*
and *LX_FloatingBox*.

--