text texture: fix alpha transparency
The alpha channel was not used properly in the 3 text textures.

It is related to the surface each on of the three textures are created from.

> LX_SolidTextTexture:

The less complicated bug to fix.

I just needed to use ```SDL_SetTextureAlphaMod()``` just after the creation of the texture
in order to get alpha transparency properly, because alpha transparency is not used.

> LX_ShadedTextTexture and LX_BlendedTextTexture

Alpha transparency is used. But it cannot be modified. So, the alpha value provided by LX_Colour
has no effect.

In order to make alpha transparency have effect on the internal texture, I had to set the blend mode.
by using ```SDL_SetSurfaceBlendMode()``` before I create the internal texture. After that, the next step is similar to
what I did with **LX_SolidTexture**.

---
Links:

- ```SDL_SetTextureAlphaMod()```: https://wiki.libsdl.org/SDL_SetTextureAlphaMod
- ```SDL_SetSurfaceBlendMode()```: https://wiki.libsdl.org/SDL_SetSurfaceBlendMode
- SDL2 and SDL_ttf text blending — forum: http://sdl.5483.n7.nabble.com/SDL2-and-SDL-ttf-text-blending-td34398.html

--