groupPlayChunk(): fix bug
> Bug

This functions always plays a sample on any channel.
It did not care about the group of channels.

> Expected behaviour

Normally, this function plays on any channel if and only if
the group specified by the tag has no channel.

Instead of doing that, it tests if the group specified by the tag has at least 1 channel
(the number of channels cannot be negative).

> Fix

Changed the condition

--